+++ baml_language/crates/baml_compiler2_visualization/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "baml_compiler2_visualization"
+version = { workspace = true }
+publish = false
+authors = { workspace = true }
+edition = { workspace = true }
+rust-version = { workspace = true }
+homepage = { workspace = true }
+documentation = { workspace = true }
+repository = { workspace = true }
+license = { workspace = true }
+
+[lib]
+doctest = false
+
+[lints]
+workspace = true
+
+[dependencies]
+baml_compiler2_ast = { workspace = true }
+indexmap = { workspace = true }
+
+[dev-dependencies]
+la-arena = { workspace = true }
