+++ baml_language/crates/baml_compiler_vir/src/control_flow.rs
@@ -1,270 +1,17 @@
 //! Control flow visualization graph for BAML functions.
 //!
-//! This module builds a graph representing the control flow structure of a
-//! function body: headers (`//#`), if/else branches, loops, match expressions,
-//! and scoped blocks. The playground uses this to visualize function execution.
-//!
-//! The graph is built by walking VIR expressions — a clean, dedicated pass
-//! separate from MIR lowering.
+//! Graph types and the AST-based builder live in `baml_compiler2_visualization`.
+//! This module re-exports those and adds a VIR-specific builder that walks
+//! the expression-only Validated IR.
 
-mod flatten;
-mod from_ast;
-
-use std::{collections::HashMap, fmt};
-
-pub use flatten::flatten_control_flow_graph;
-pub use from_ast::build_control_flow_graph_from_ast;
-use indexmap::IndexMap;
+// Re-export everything from the visualization crate so existing consumers
+// (db.rs, baml_tests, etc.) continue to work via `baml_compiler_vir::control_flow::*`.
+pub use baml_compiler2_visualization::control_flow::*;
 
 use crate::{BinaryOp, Expr, ExprBody, ExprId, Literal, MatchArm, Pattern, UnaryOp};
 
 // ---------------------------------------------------------------------------
-// Data structures (mirror engine's control_flow.rs:12-120)
-// ---------------------------------------------------------------------------
-
-/// Opaque node identifier.
-#[derive(Copy, Clone, Debug, PartialEq, Eq, Hash)]
-pub struct NodeId(u32);
-
-impl NodeId {
-    fn new(raw: u32) -> Self {
-        Self(raw)
-    }
-
-    pub fn raw(&self) -> u32 {
-        self.0
-    }
-}
-
-impl fmt::Display for NodeId {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", self.0)
-    }
-}
-
-/// Segment of a log-filter key path.
-#[derive(Clone, Debug, PartialEq, Eq, Hash)]
-enum PathSegment {
-    FunctionRoot { ordinal: u16 },
-    Header { slug: String, ordinal: u16 },
-    BranchGroup { slug: String, ordinal: u16 },
-    BranchArm { slug: String, ordinal: u16 },
-    Loop { slug: String, ordinal: u16 },
-    OtherScope { slug: String, ordinal: u16 },
-}
-
-/// The type of a visualization node.
-#[derive(Clone, Debug)]
-pub enum NodeType {
-    FunctionRoot,
-    HeaderContextEnter,
-    BranchGroup,
-    BranchArm,
-    Loop,
-    OtherScope,
-}
-
-/// A node in the control flow visualization graph.
-#[derive(Clone, Debug)]
-pub struct Node {
-    pub id: NodeId,
-    pub parent_node_id: Option<NodeId>,
-    pub log_filter_key: String,
-    pub label: String,
-    /// Position-independent reference into the VIR expression arena.
-    /// Callers resolve to `Span` at rendering time via `ExprBody.source_spans[expr_id]`.
-    pub source_expr: Option<ExprId>,
-    pub node_type: NodeType,
-}
-
-impl Node {
-    fn new(
-        id: NodeId,
-        parent_node_id: Option<NodeId>,
-        log_filter_key: impl Into<String>,
-        label: impl Into<String>,
-        source_expr: Option<ExprId>,
-        node_type: NodeType,
-    ) -> Self {
-        Self {
-            id,
-            parent_node_id,
-            log_filter_key: log_filter_key.into(),
-            label: label.into(),
-            source_expr,
-            node_type,
-        }
-    }
-
-    fn root(id: NodeId, log_filter_key: impl Into<String>, label: impl Into<String>) -> Self {
-        Self::new(
-            id,
-            None,
-            log_filter_key,
-            label,
-            None,
-            NodeType::FunctionRoot,
-        )
-    }
-}
-
-/// A directed edge in the visualization graph.
-#[derive(Clone, Debug)]
-pub struct Edge {
-    pub src: NodeId,
-    pub dst: NodeId,
-}
-
-/// The control flow visualization graph.
-#[derive(Clone, Debug, Default)]
-pub struct ControlFlowGraph {
-    pub nodes: IndexMap<NodeId, Node>,
-    pub edges_by_src: IndexMap<NodeId, Vec<Edge>>,
-}
-
-// ---------------------------------------------------------------------------
-// Graph builder accumulator
-// ---------------------------------------------------------------------------
-
-struct GraphAccumulator {
-    nodes: IndexMap<NodeId, Node>,
-    edges: Vec<Edge>,
-    next_node_id: u32,
-}
-
-impl Default for GraphAccumulator {
-    fn default() -> Self {
-        Self {
-            nodes: IndexMap::new(),
-            edges: Vec::new(),
-            next_node_id: 0,
-        }
-    }
-}
-
-impl GraphAccumulator {
-    fn allocate_id(&mut self) -> NodeId {
-        let id = NodeId::new(self.next_node_id);
-        self.next_node_id += 1;
-        id
-    }
-
-    fn add_node(&mut self, node: Node) {
-        self.nodes.insert(node.id, node);
-    }
-
-    fn add_edge(&mut self, src: NodeId, dst: NodeId) {
-        self.edges.push(Edge { src, dst });
-    }
-
-    fn finish(self) -> ControlFlowGraph {
-        let mut edges_by_src: IndexMap<NodeId, Vec<Edge>> = IndexMap::new();
-        for edge in self.edges {
-            edges_by_src.entry(edge.src).or_default().push(edge);
-        }
-        ControlFlowGraph {
-            nodes: self.nodes,
-            edges_by_src,
-        }
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Frame stack (scope tracking during traversal)
-// ---------------------------------------------------------------------------
-
-#[derive(Clone, Default)]
-struct FrameCounters {
-    header: u16,
-    branch_group: u16,
-    branch_arm: u16,
-    loop_node: u16,
-    other_scope: u16,
-}
-
-enum CounterKind {
-    Header,
-    BranchGroup,
-    BranchArm,
-    Loop,
-    OtherScope,
-}
-
-impl FrameCounters {
-    fn next(&mut self, kind: &CounterKind) -> u16 {
-        match kind {
-            CounterKind::Header => {
-                let c = self.header;
-                self.header += 1;
-                c
-            }
-            CounterKind::BranchGroup => {
-                let c = self.branch_group;
-                self.branch_group += 1;
-                c
-            }
-            CounterKind::BranchArm => {
-                let c = self.branch_arm;
-                self.branch_arm += 1;
-                c
-            }
-            CounterKind::Loop => {
-                let c = self.loop_node;
-                self.loop_node += 1;
-                c
-            }
-            CounterKind::OtherScope => {
-                let c = self.other_scope;
-                self.other_scope += 1;
-                c
-            }
-        }
-    }
-}
-
-enum FrameEntry {
-    FunctionRoot,
-    Header { level: u8 },
-    BranchGroup,
-    BranchArm,
-    Loop,
-    OtherScope,
-}
-
-impl FrameEntry {
-    // The types of children that, if peers in a code sequence,
-    // are "linear" (follow one another).
-    fn children_are_linear(&self) -> bool {
-        !matches!(self, FrameEntry::BranchGroup)
-    }
-}
-
-struct Frame {
-    entry: FrameEntry,
-    node_id: NodeId,
-    lexical_segment: Option<PathSegment>,
-    counters: FrameCounters,
-    last_linear_child: Option<NodeId>,
-}
-
-impl Frame {
-    fn new(entry: FrameEntry, node_id: NodeId, lexical_segment: Option<PathSegment>) -> Self {
-        Self {
-            entry,
-            node_id,
-            lexical_segment,
-            counters: FrameCounters::default(),
-            last_linear_child: None,
-        }
-    }
-
-    fn next_ordinal(&mut self, kind: &CounterKind) -> u16 {
-        self.counters.next(kind)
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Public API
+// Public API — VIR-specific
 // ---------------------------------------------------------------------------
 
 /// Build a control flow visualization graph from a VIR expression body.
@@ -272,34 +19,6 @@ pub fn build_control_flow_graph(function_name: &str, body: &ExprBody) -> Control
     let mut builder = GraphBuilder::new(function_name, body);
     builder.visit_expr(body.root);
     builder.finish()
-}
-
-/// Build a simple 2-node control flow graph for an LLM function.
-///
-/// Produces: `FunctionRoot` -> `OtherScope("LLM client: <client_name>")`.
-pub fn build_llm_control_flow_graph(function_name: &str, client_name: &str) -> ControlFlowGraph {
-    let mut graph = GraphAccumulator::default();
-    let root_id = graph.allocate_id();
-    let root_segment = PathSegment::FunctionRoot { ordinal: 0 };
-    let root_key = encode_segments(function_name, std::slice::from_ref(&root_segment));
-    graph.add_node(Node::root(root_id, root_key, function_name));
-
-    let slug = slug_or_default("llm", "llm");
-    let segment = PathSegment::OtherScope { slug, ordinal: 0 };
-    let log_filter_key = encode_segments(function_name, &[root_segment, segment]);
-    let scope_id = graph.allocate_id();
-    let node = Node::new(
-        scope_id,
-        Some(root_id),
-        log_filter_key,
-        format!("LLM client: {client_name}"),
-        None,
-        NodeType::OtherScope,
-    );
-    graph.add_node(node);
-    graph.add_edge(root_id, scope_id);
-
-    graph.finish()
 }
 
 // ---------------------------------------------------------------------------
@@ -383,7 +102,11 @@ impl<'a> GraphBuilder<'a> {
                 if needs_scope {
                     let pat_name = self.format_pattern(*pattern);
                     let label = format!("let {pat_name} = ...");
-                    self.emit_other_scope(*value, Some(label), Some(id));
+                    self.emit_other_scope(
+                        *value,
+                        Some(label),
+                        Some(id.into_raw().into_u32()),
+                    );
                 } else {
                     self.visit_expr(*value);
                 }
@@ -391,7 +114,7 @@ impl<'a> GraphBuilder<'a> {
             }
 
             Expr::NotifyBlock { name, level } => {
-                self.enter_header(name.as_ref(), *level, Some(id));
+                self.enter_header(name.as_ref(), *level, Some(id.into_raw().into_u32()));
             }
 
             Expr::If {
@@ -452,7 +175,7 @@ impl<'a> GraphBuilder<'a> {
             parent_id,
             log_filter_key,
             label,
-            Some(expr_id),
+            Some(expr_id.into_raw().into_u32()),
             NodeType::BranchGroup,
         );
         self.graph.add_node(node);
@@ -463,7 +186,11 @@ impl<'a> GraphBuilder<'a> {
 
         // First arm: "if (condition)"
         let arm_label = format!("if ({})", render_expr_compact(self.body, condition));
-        self.visit_branch_arm(arm_label, then_branch, Some(then_branch));
+        self.visit_branch_arm(
+            arm_label,
+            then_branch,
+            Some(then_branch.into_raw().into_u32()),
+        );
 
         // Flatten else-if chains
         let mut current_else = else_branch;
@@ -477,11 +204,19 @@ impl<'a> GraphBuilder<'a> {
                 } => {
                     let arm_label =
                         format!("else if ({})", render_expr_compact(self.body, else_cond));
-                    self.visit_branch_arm(arm_label, else_then, Some(else_then));
+                    self.visit_branch_arm(
+                        arm_label,
+                        else_then,
+                        Some(else_then.into_raw().into_u32()),
+                    );
                     current_else = else_else;
                 }
                 _ => {
-                    self.visit_branch_arm("else".to_string(), else_id, Some(else_id));
+                    self.visit_branch_arm(
+                        "else".to_string(),
+                        else_id,
+                        Some(else_id.into_raw().into_u32()),
+                    );
                     current_else = None;
                 }
             }
@@ -489,13 +224,21 @@ impl<'a> GraphBuilder<'a> {
 
         // Synthetic "else" arm if no else branch
         if else_branch.is_none() {
-            self.emit_synthetic_branch_arm("else".to_string(), Some(expr_id));
+            self.emit_synthetic_branch_arm(
+                "else".to_string(),
+                Some(expr_id.into_raw().into_u32()),
+            );
         }
 
         self.pop_frames_to(parent_depth);
     }
 
-    fn visit_branch_arm(&mut self, label: String, body_expr: ExprId, source_expr: Option<ExprId>) {
+    fn visit_branch_arm(
+        &mut self,
+        label: String,
+        body_expr: ExprId,
+        source_expr: Option<u32>,
+    ) {
         let parent_depth = self.frames.len();
         let ordinal = {
             let frame = self
@@ -532,7 +275,7 @@ impl<'a> GraphBuilder<'a> {
     }
 
     /// Create a branch arm node with no body traversal (for synthetic else arms).
-    fn emit_synthetic_branch_arm(&mut self, label: String, source_expr: Option<ExprId>) {
+    fn emit_synthetic_branch_arm(&mut self, label: String, source_expr: Option<u32>) {
         let ordinal = {
             let frame = self
                 .frames
@@ -593,7 +336,7 @@ impl<'a> GraphBuilder<'a> {
             parent_id,
             log_filter_key,
             label,
-            Some(expr_id),
+            Some(expr_id.into_raw().into_u32()),
             NodeType::BranchGroup,
         );
         self.graph.add_node(node);
@@ -604,7 +347,11 @@ impl<'a> GraphBuilder<'a> {
 
         for arm in arms {
             let arm_label = self.format_pattern(arm.pattern);
-            self.visit_branch_arm(arm_label, arm.body, Some(arm.body));
+            self.visit_branch_arm(
+                arm_label,
+                arm.body,
+                Some(arm.body.into_raw().into_u32()),
+            );
         }
 
         self.pop_frames_to(parent_depth);
@@ -637,7 +384,7 @@ impl<'a> GraphBuilder<'a> {
             parent_id,
             log_filter_key,
             label,
-            Some(expr_id),
+            Some(expr_id.into_raw().into_u32()),
             NodeType::Loop,
         );
         self.graph.add_node(node);
@@ -652,7 +399,7 @@ impl<'a> GraphBuilder<'a> {
     // -- Headers --
 
     #[allow(clippy::cast_possible_truncation)]
-    fn enter_header(&mut self, title: &str, level: usize, source_expr: Option<ExprId>) {
+    fn enter_header(&mut self, title: &str, level: usize, source_expr: Option<u32>) {
         let level = (level as u8).max(1);
         self.pop_headers_to_level(level - 1);
 
@@ -698,7 +445,7 @@ impl<'a> GraphBuilder<'a> {
         &mut self,
         inner_expr: ExprId,
         label: Option<String>,
-        source_expr: Option<ExprId>,
+        source_expr: Option<u32>,
     ) {
         let parent_depth = self.frames.len();
         let ordinal = {
@@ -783,7 +530,7 @@ impl<'a> GraphBuilder<'a> {
 }
 
 // ---------------------------------------------------------------------------
-// Compact expression renderer (for labels)
+// VIR-specific compact expression renderer (for labels)
 // ---------------------------------------------------------------------------
 
 fn render_expr_compact(body: &ExprBody, id: ExprId) -> String {
@@ -861,147 +608,13 @@ fn format_literal(lit: &Literal) -> String {
 }
 
 // ---------------------------------------------------------------------------
-// Utility functions (ported from engine)
-// ---------------------------------------------------------------------------
-
-fn slugify(input: &str) -> String {
-    let mut slug = String::with_capacity(input.len());
-    let mut last_dash = false;
-    for ch in input.chars() {
-        if ch.is_ascii_alphanumeric() {
-            slug.push(ch.to_ascii_lowercase());
-            last_dash = false;
-        } else if !last_dash {
-            slug.push('-');
-            last_dash = true;
-        }
-    }
-    slug.trim_matches('-').to_string()
-}
-
-fn slug_or_default(label: &str, default: &str) -> String {
-    let candidate = slugify(label);
-    if candidate.is_empty() {
-        default.to_string()
-    } else {
-        candidate
-    }
-}
-
-pub fn describe_node_type(node_type: &NodeType) -> &'static str {
-    match node_type {
-        NodeType::FunctionRoot => "function",
-        NodeType::HeaderContextEnter => "header",
-        NodeType::BranchGroup => "branch-group",
-        NodeType::BranchArm => "branch-arm",
-        NodeType::Loop => "loop",
-        NodeType::OtherScope => "other-scope",
-    }
-}
-
-fn encode_segments(function: &str, segments: &[PathSegment]) -> String {
-    let mut encoded = String::from(function);
-    for segment in segments {
-        encoded.push('|');
-        match segment {
-            PathSegment::FunctionRoot { ordinal } => {
-                encoded.push_str("root:");
-                encoded.push_str(&ordinal.to_string());
-            }
-            PathSegment::Header { slug, ordinal } => {
-                encoded.push_str("hdr:");
-                encoded.push_str(slug);
-                encoded.push(':');
-                encoded.push_str(&ordinal.to_string());
-            }
-            PathSegment::BranchGroup { slug, ordinal } => {
-                encoded.push_str("bg:");
-                encoded.push_str(slug);
-                encoded.push(':');
-                encoded.push_str(&ordinal.to_string());
-            }
-            PathSegment::BranchArm { slug, ordinal } => {
-                encoded.push_str("arm:");
-                encoded.push_str(slug);
-                encoded.push(':');
-                encoded.push_str(&ordinal.to_string());
-            }
-            PathSegment::Loop { slug, ordinal } => {
-                encoded.push_str("loop:");
-                encoded.push_str(slug);
-                encoded.push(':');
-                encoded.push_str(&ordinal.to_string());
-            }
-            PathSegment::OtherScope { slug, ordinal } => {
-                encoded.push_str("scope:");
-                encoded.push_str(slug);
-                encoded.push(':');
-                encoded.push_str(&ordinal.to_string());
-            }
-        }
-    }
-    encoded
-}
-
-// ---------------------------------------------------------------------------
-// Helpers used by flatten module
-// ---------------------------------------------------------------------------
-
-pub(crate) fn build_children_map(nodes: &IndexMap<NodeId, Node>) -> HashMap<NodeId, Vec<NodeId>> {
-    let mut children: HashMap<NodeId, Vec<NodeId>> = HashMap::new();
-    for node in nodes.values() {
-        if let Some(parent) = node.parent_node_id {
-            children.entry(parent).or_default().push(node.id);
-        }
-    }
-    children
-}
-
-pub(crate) fn node_depth(node_id: NodeId, nodes: &IndexMap<NodeId, Node>) -> usize {
-    let mut depth = 0;
-    let mut current = Some(node_id);
-    while let Some(id) = current {
-        depth += 1;
-        current = nodes.get(&id).and_then(|node| node.parent_node_id);
-    }
-    depth
-}
-
-// ---------------------------------------------------------------------------
-// Display for ControlFlowGraph (for snapshot tests)
-// ---------------------------------------------------------------------------
-
-impl fmt::Display for ControlFlowGraph {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        writeln!(f, "Nodes:")?;
-        for (id, node) in &self.nodes {
-            let parent = node
-                .parent_node_id
-                .map(|p| p.to_string())
-                .unwrap_or_else(|| "-".to_string());
-            writeln!(
-                f,
-                "  [{id}] parent={parent} type={} label={:?}",
-                describe_node_type(&node.node_type),
-                node.label
-            )?;
-        }
-        writeln!(f, "Edges:")?;
-        for edges in self.edges_by_src.values() {
-            for edge in edges {
-                writeln!(f, "  {} -> {}", edge.src, edge.dst)?;
-            }
-        }
-        Ok(())
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Tests
+// Tests (VIR-specific)
 // ---------------------------------------------------------------------------
 
 #[cfg(test)]
 mod tests {
+    use std::collections::HashMap;
+
     use la_arena::Arena;
 
     use super::*;
